+++ pEcho/include/rdr_obj.hpp
@@ -14,12 +14,12 @@ class rdr_obj : public scenario {
 public:
 	// Management Functions
 	void setupDefaultScenario();
-  void createSuperstructure(int p_soBodies, double p_cMass, double p_oMass, double p_cRadius, double p_oRadius, double p_cPosX, double p_cPosY, double p_cVelX, double p_cVelY, double p_coSpacing, double p_sRadius, const float p_Color[3]);
+  void createSuperstructure(int p_soBodies, float p_cMass, float p_oMass, float p_cRadius, float p_oRadius, float p_cPosX, float p_cPosY, float p_cVelX, float p_cVelY, float p_coSpacing, float p_sRadius, const float p_Color[3]);
   void updateSharedArea(sharedStage* l_sharedDataAccess);
   void updateLocalStore(sharedStage* l_sharedDataAccess);
   // Render-Specific Functions Here
   void drawBody(int bodyID);
   void drawScene(void);
   // Check Coordinates
-  int checkCoord(double x, double y);
+  int checkCoord(float x, float y);
 };
