+++ pEcho/include/scenario.hpp
@@ -13,8 +13,8 @@ protected:
   com::bodyVector bodyStore; // List, C++ Std library vector, dynamic array.
 
   // Simulation Physical Constants (Can be modified)
-  double UGC;
-  double IDT;
+  float UGC;
+  float IDT;
   int IPF;
 
   bool scenarioChanged = true;
@@ -25,16 +25,16 @@ public:
   scenario();
   ~scenario();
 
-  void newBody(double p_Mass, double p_Radius, double p_PosX, double p_PosY, double p_VelX, double p_VelY, const float p_Color[3]);
+  void newBody(float p_Mass, float p_Radius, float p_PosX, float p_PosY, float p_VelX, float p_VelY, const float p_Color[3]);
   void delBody(int index);
   void delAllBodies(int index);
 
   void updateLocalControl(sharedStage* l_sharedDataAccess);
-  void updateLocalControl(double p_UGC, double p_IDT, int p_IPF);
+  void updateLocalControl(float p_UGC, float p_IDT, int p_IPF);
 
   void updateSharedControl(sharedStage* l_sharedDataAccess);
 
-  double getUGC();
-  double getIDT();
+  float getUGC();
+  float getIDT();
   int getIPF();
 };
